Fix sync function so that it doesn't remove nodes/envs wrong
'dos.py sync' function was using a bad call to libvirt which only
returns stopped domains. This causes envs with nodes on to be pruned.
If all nodes are on, then the whole env is removed from db

It further removes all nodes that are not in the db, this is not OK
 because it will undefine nodes that might have never been managed
 by devops. This is disabled untill a safer option is designed.

Change-Id: Ic257d6bd98d0a9b38df4ddb7aab04517762302a4
Closes-bug: #1316856