Change nodes so that they end in admin mac suffix instead of number
Previously the nodes would end in a number iterator to ensure they were
unique, however this makes working with the nodes generated this way
difficult as they are referred to by their admin interface mac.

This changes it so the slave nodes will receive the suffix of the admin
mac.

Closes-bug: #1424276
Change-Id: Iab5b5caa1b10c89d32eb60a1f5b4cbfbb4e58d56