Configure VMs to use host's clock source, simplify timesync
Correct time syncronization is not possible from devops now because
it requires close interaction with the OpenStack clusters (because
of p_ntp resources in the pacemaker) that is out of scope of devops
purposes.

Instead of making too complicated time syncronization routines in
devops, let's use the host clock source as the single source for
all VMs for quick setting the actual time for nodes.

System tests are still allowed to use more suitable ways for time
syncronization.

Change-Id: I6d3b2a89cd72b24d6c4f558534cd9dea82925b88
Related-Bug: #1433484