Fix node attach_to_networks for slave interfaces
The node.attach_to_networks function has a bug where the get_networks
call is not actually returning the available networks for the node.
This causes the slave nodes to be created with out network interfaces.
To fix this, the get_networks call is updated with the name__in
parameter to properly perform an IN sql query since the get_networks
call just passes all paramters on to a queryset.

Change-Id: Ia16c0206c1c6d8817970c6d8c9744f6885adba4f
Closes-Bug: 1444058