Initial non-eth0 admin interface
This change adds flexible bootstrap of
a nodes when admin network ethernet
devices are not eth0. Admin node
should not be booted with bootmenu
because it hangs in bootmenu.

That only checks networks ordering
and add xml attribute 'bootmenu'
to virtual machine xml description.
It checks that node is admin or slave.
If node is a master, bootmenu should be
enabled only when there is a usb boot.
If node is a slave, bootmenu should be
enabled if boot interface on any of
nodegroup is not on eth0.

pxe_boot_interface_is_eth0 returns false
if node is admin because we do not boot
admin node on pxe.
if settings.MULTIPLE_NETWORKS is false
network is set to none because we do not
create networks for second nodegroup
when tests do not check MULTIPLE_NETWORKS.
That part is tricky and will be refactored
when network settings are being cleaned up.

Partial-Bug: #1439417
Change-Id: I87db9734418347161f60b854c1619c253ac9ab97