Fix revert-resume Authentication Error
When we get private keys from the master node
there are two keys '/root/.ssh/id_rsa' and
 '/root/.ssh/bootstrap.rsa'.
An error occures when we try to use a wrong key,
but on the next iteration we use right one
and authentication passes.
So it's just cosmetic defect and patch fixes an exception
mechanism.

Change-Id: If167abc34f1c4060ab3b791204dc7245d06805d9
Closes-Bug: 1494856