[WIP] Add partial support for external snapshots
This change add partial support for external snapshots from libvirt.
Allow:
- create disk and memory snapshot in separated file
- restore domain state
- delete not used external snapshot files

To use external snapshots libvirt must be in version >= 1.2.12

This change does not add full support for external snapshots and it
not replace internal snapshots functionality. It goal is to
create separate snapshots files, which could be reused on other
systems, without need to copy one big qcow file with all snapshots.

Change-Id: I415d753a0847f4c6de2b4b0bb79aa5ead37e5d8d