Fuel-devops dos.py show env occurs exception
The Environment.DoesNotExist exception will be generated
because django will not found corresponding model.
shall.py will be invoked by dos.py and
shall.py module gets 'env' by name which was passed through CI
where 'env' object shall be taken from settings
stored in database before.
If requested 'env' is not presented in database then
django lib throw exception but shall.py module
don't intercept it and it is root cause.

The Environment.DoesNotExist shall be intercepted
in shall.py in init function and reasonable message
shall be shown in this case.
We will be blocked to continue work with odd 'env'
which is not presented in database because 'env' is
required for most of shall.py public methods.
in this case we have to return "-1" to CI and let to know
another processes that we was failed.

Closes-Bug: #1515258

Change-Id: I15c3e47867c5d4c168ae510110c46acde8933841