Fix devops create method
This change fixes the regression introduced by change
Iad081bee94c4d3b404eda57b7c1c31011b837ba7 which broke the use of the
'dos.py create' shell command. The refactoring of the Environment to
support templates changed when the nodes and networks get defined and
saved in the DB. Since the ordering changed, this creation of the admin
node via the do_create method no longer works. Additionaly the
Environment.create_interfaces function was changed but not updated in
all places where it was used. This change adds the previous verison of
the function back as create_interfaces_from_networks.

This change ensures the networks are defined prior to the creation of
the admin node for the create command, slave nodes are created as part
of the do_create function and updates the node model so that it can
properly be attached to networks.

Change-Id: I965f647285d8eca28a71a2d31708eee7a6d36b4d
Closes-Bug: #1521520