Fix devops create method
This change fixes the regression introduced by change
Iad081bee94c4d3b404eda57b7c1c31011b837ba7 which broke the use of the
'dos.py create' shell command. The refactoring of the Environment to
support templates removed the networks from being properly defined via
the create method which resulted in UUIDs not being properly generated
for the networks. Additionaly the Environment.create_interfaces
function was changed but not updated in all places where it was used.

This change ensures the networks are defined prior to the creation of
the admin node for the create command and updates the node model so
that it can properly be attached to networks.

Change-Id: I965f647285d8eca28a71a2d31708eee7a6d36b4d
Closes-Bug: #1521520