Add admin-setup-centos7 and iface support
The anaconda boot parameters have changed with centos7 so the networking
information needs to be provided as part of the ip parameter. The gw,
dns1, and netmask parameters are no longer supported. A new
admin-setup-centos7 has been added to support the configuration of
centos7 master nodes. Additionall the --iface option has been added for
centos7 to support the use of predictable interface names and the old
eth## style. The default for centos7 is predictable interface names.

Change-Id: I91b7817bac2ef0c3c558b3b337c2661df829c7e6
Closes-Bug: #1524440