Add centos7 support to admin-setup
The anaconda boot parameters have changed with centos7 so the networking
information needs to be provided as part of the ip parameter. The gw,
dns1, and netmask parameters are no longer supported. As such a new
configuration option has been added to admin-setup to indicate what
version of centos is being setup. The default is 7 as all newer fuel
builds use centos7 as default. To support previous centos6 releases, you
can use --admin-centos-version=6 as part of the admin-setup command.

Change-Id: I91b7817bac2ef0c3c558b3b337c2661df829c7e6
Closes-Bug: #1524440