Fix slave IP address getting
We use a MAC address to match devops and nailgun nodes.
Sometimes the MAC address of the br-fw-admin bridge
on an already deployed environment has been changed to
the OVS interface MAC.
In that cases we couldn't get an IP by a new MAC,
but the old MAC is stored as a part of the name of the node.
Fix changes expression that is used for getting IP.

Change-Id: Ia25b03bd970a9e8da19c2cb4356552c177fc9244
Closes-Bug:1542275