Fix slave IP address getting
We use a MAC address to match devops and nailgun nodes.
Sometimes the MAC address of the br-fw-admin bridge
on an already deployed environment has been changed
to the OVS interface MAC.
In that cases we couldn't get an IP by a new MAC,
but the old MAC is stored as a part of the name of the node.
Fix changes expression that is used for getting IP.

Change-Id: Icf1ff0a85f3a5555d1b7d70a652b2aba9b8f5606
Closes-Bug:1542275