Add dict wrapper
Dict wrapper is used for string format method using
postponed keys resolving.
With this class, format will not fail in case if key not found,
and leaves the unknown key in the string 'as is'.
This behaviour is required as a workaround for strings with
bash-like variables like ${SOMEVAR} to not fail the format because
of 'unknown key: SOMEVAR'.
Returning the 'format' instead of 'replace' is required to use
the escaping of the '{{' '}}' in the string to get '{' '}'.

Change-Id: I1c2bbb083a544eb366c7b00eb1002cc4da652c33